requirements.txt
asgiref==3.4.1
beautifulsoup4==4.9.3
bootstrap4==0.1.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
Django==3.2.5
django-bootstrap4==3.0.1
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
docopt==0.6.2
gunicorn==20.1.0
httpie==2.4.0
idna==3.2
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
openapi-codec==1.3.2
Pillow==8.3.1
pipreqs==0.4.10
psycopg2==2.9.1
pycparser==2.20
Pygments==2.10.0
PySocks==1.7.1
pytz==2021.1
requests==2.26.0
requests-toolbelt==0.9.1
simplejson==3.17.5
soupsieve==2.2.1
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.3.0
yarg==0.1.9
